Use MYVIMRC env var instead of VIM in example
The current example in the README sets the VIM environment variable. The problem
with that approach is that vim will then assume that all of the vim-related
files (basically, the contents of, say, /usr/share/vim/vim73 on Ubuntu 12.04)
are in this directory.

This means that you won't be able to enable syntax highlighting, because you'll
get the error:

```
Error detected while processing /tmp/tmp.aPn9BMrJAw/.sshrc.d/vimrc:
line    3:
E484: Can't open file /tmp/tmp.aPn9BMrJAw/.sshrc.d/syntax/syntax.vim
```

This tripped me up when I hit it. I tried copying the syntax.vim file to the
appropriate place, but got another error because some other files weren't there.

A simpler approach is to just set the MYVIMRC environment variable, so that vim
uses the custom vimrc file, but uses the local system files for the rest of the
stuff.